Added sortByPt method which is meant to be used as a predicate for the standard library sort algorithm. I saw that it was defined in multiple places so this is where it will be defined centrally.
Added Jmu's DorkyEventIdentifier though in a much cleaner implementation. Again
it was placed here so that it's defined centrally. To use, make an instance of
DorkyEventIdentifier, e.g. DorkyEventIdentifier dei, outside of your event loop
and then do bool isduplicate = dei.is_duplicate(DorkyEvent()). I left it in more
or less the same form as I found it, but clearly we could change this to just be
dei.is_duplicate().

I moved all this stuff here because it needed to go somewhere. Let's keep an eye
on utilities* and be open to changes so that we don't run into the same problems
here that we ran into with selections*